+++ demos/world/src/main/java/is/codion/framework/demos/world/ui/ContinentPanel.java
@@ -7,6 +7,7 @@ import is.codion.framework.demos.world.domain.api.World.Country;
 import is.codion.framework.demos.world.model.ContinentModel;
 import is.codion.swing.framework.model.SwingEntityModel;
 import is.codion.swing.framework.ui.EntityPanel;
+import is.codion.swing.framework.ui.EntityTablePanel;
 
 import org.jfree.chart.ChartPanel;
 
@@ -29,8 +30,9 @@ final class ContinentPanel extends EntityPanel {
   ContinentPanel(SwingEntityModel continentModel) {
     super(continentModel, new ContinentTablePanel(continentModel.tableModel()),
             new ContinentPanelLayout());
-    countryPanel = new EntityPanel(continentModel.detailModel(Country.TYPE));
-    countryPanel.tablePanel().configure().includeConditionPanel(false);
+    countryPanel = new EntityPanel(continentModel.detailModel(Country.TYPE),
+            new EntityTablePanel(continentModel.tableModel(),
+                    settings -> settings.includeConditionPanel(false)));
   }
 
   private static final class ContinentPanelLayout implements PanelLayout {