+++ demos/world/src/main/java/is/codion/framework/demos/world/ui/CountryTablePanel.java
@@ -21,8 +21,9 @@ import java.awt.Dimension;
 final class CountryTablePanel extends EntityTablePanel {
 
   CountryTablePanel(SwingEntityTableModel tableModel) {
-    super(tableModel);
-    configure().editableAttributes().remove(Country.CAPITAL_FK);
+    super(tableModel, settings ->
+            settings.editableAttributes(attributes ->
+                    attributes.remove(Country.CAPITAL_FK)));
   }
 
   @Override