+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityDependenciesPanel.java
@@ -117,7 +117,9 @@ public final class EntityDependenciesPanel extends JPanel {
   }
 
   private static EntityTablePanel createTablePanel(Collection<Entity> entities, EntityConnectionProvider connectionProvider) {
-    EntityTablePanel tablePanel = new EntityTablePanel(SwingEntityTableModel.tableModel(entities, connectionProvider)) {
+    SwingEntityTableModel tableModel = SwingEntityTableModel.tableModel(entities, connectionProvider);
+    EntityTablePanel tablePanel = new EntityTablePanel(tableModel, settings -> settings
+            .includeConditionPanel(false)) {
       @Override
       protected Controls createPopupMenuControls(List<Controls> additionalPopupMenuControls) {
         Controls popupMenuControls = Controls.controls();
@@ -133,7 +135,6 @@ public final class EntityDependenciesPanel extends JPanel {
         return popupMenuControls;
       }
     };
-    tablePanel.configure().includeConditionPanel(false);
 
     return tablePanel.initialize();
   }