+++ changelog.md
@@ -22,6 +22,7 @@ Codion Change Log
 - EntityComponents, range validation no longer added to number fields by default.
 - EntityDialogs.DefaultEntityEditDialogBuilder.InputValidator now also uses the entity validator when validating input.
 - EntityTablePanel refactored.
+- EntityTablePanel, settings now configured in a constructor lambda.
 
 ## 0.17.31
 ### is.codion.common